[WIP] Add theme support
This also serves as a nice helping hand with development, because I no longer have to guess which is my primary text color, since I now have a standardized system. During the next few commits, it will be my focus to bring all the subpages and components up to speed.